Begin Refactor of Soda Machine Transaction() to get value of coins inserted
Began refactoring the Soda Machine Transaction method to now accept a precalculated sum of the coins. Now instead of calculating it when needed, it will be passed down. It will then be used as the value for the amount of change to give, either if Transaction failed or after price of soda has been subtracted and normal change is returned.