Major Refactor of Soda Machine Class with documenting comments.
Major refactor of the Soda Machine class.
Will no longer pass a reference of a list of coins or an instance of a soda class through the validation process. Only the amount of the coins inserted and the name of the soda selection.
Refactored all methods to fit this new variable flow.
Added new comments to all cosntructors and methods.
Deleted dead/Depreciated code, with exception of methods that may be implemented at a later date.